Rearranging to straighten out circular import issues
